@@ -10,6 +10,7 @@
         "ext-iconv": "*",
         "ext-intl": "*",
         "ext-json": "*",
+        "ext-soap": "*",
         "composer/package-versions-deprecated": "^1.8",
         "craue/twigextensions-bundle": "^2.4",
         "debril/rss-atom-bundle": "^5.0",
@@ -51,13 +52,13 @@
         "symfony/http-client": "4.4.*",
         "symfony/ldap": "4.4.*",
         "symfony/lock": "4.4.*",
+        "symfony/mailer": "4.4.*",
         "symfony/monolog-bundle": "^3.3",
         "symfony/polyfill-apcu": "^1.10",
         "symfony/process": "4.4.*",
         "symfony/security-bundle": "4.4.*",
         "symfony/serializer": "4.4.*",
         "symfony/stopwatch": "4.4.*",
-        "symfony/swiftmailer-bundle": "^3.2",
         "symfony/translation": "4.4.*",
         "symfony/twig-bundle": "4.4.*",
         "symfony/validator": "4.4.*",
