@@ -27,6 +27,9 @@
  */
 include_once('classes/cs_room_item.php');
 
+use App\Mail\Mailer;
+use Symfony\Component\Mime\Address;
+use Symfony\Component\Mime\Email;
 use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
 /** group room
@@ -1142,23 +1145,23 @@ class cs_grouproom_item extends cs_room_item {
          }
 
          // send email
-         $emailFrom = $symfonyContainer->getParameter('commsy.email.from');
          $fromName = $translator->getMessage('SYSTEM_MAIL_MESSAGE',$current_portal->getTitle());
 
-         $message = (new \Swift_Message())
-            ->setSubject($subject)
-            ->setBody($body, 'text/plain')
-            ->setFrom([$emailFrom => $fromName])
-            ->setTo($value);
+         $message = (new Email())
+             ->subject($subject)
+             ->html(nl2br($body))
+             ->to($value);
 
          if ($current_user) {
             $email = $current_user->getEmail();
             if (!empty($email)) {
-               $message->setReplyTo([$email => $current_user->getFullname()]);
+               $message->replyTo(new Address($email, $current_user->getFullName()));
             }
          }
 
-         $symfonyContainer->get('mailer')->send($message);
+         /** @var Mailer $mailer */
+         $mailer = $symfonyContainer->get(Mailer::class);
+         $mailer->sendEmailObject($message, $fromName);
 
          $translator->setSelectedLanguage($save_language);
          unset($save_language);
