@@ -2,7 +2,6 @@
 
 namespace App\Controller;
 
-use App\Entity\Portal;
 use App\Entity\Room;
 use App\Entity\User;
 use App\Entity\ZzzRoom;
@@ -11,6 +10,8 @@ use App\Filter\HomeFilterType;
 use App\Filter\RoomFilterType;
 use App\Form\Type\ContextType;
 use App\Form\Type\ModerationSupportType;
+use App\Mail\Mailer;
+use App\Mail\RecipientFactory;
 use App\Repository\PortalRepository;
 use App\Repository\UserRepository;
 use App\Room\Copy\LegacyCopy;
@@ -28,7 +29,6 @@ use DateTimeImmutable;
 use Exception;
 use Lexik\Bundle\FormFilterBundle\Filter\FilterBuilderUpdater;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -37,6 +37,7 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
@@ -47,13 +48,16 @@ use Symfony\Contracts\Translation\TranslatorInterface;
  */
 class RoomController extends AbstractController
 {
-    private \Swift_Mailer $mailer;
+    /**
+     * @var Mailer
+     */
+    private Mailer $mailer;
 
     /**
      * @required
-     * @param \Swift_Mailer $mailer
+     * @param Mailer $mailer
      */
-    public function setMailer(\Swift_Mailer $mailer)
+    public function setMailer(Mailer $mailer)
     {
         $this->mailer = $mailer;
     }
@@ -337,30 +341,19 @@ class RoomController extends AbstractController
             $data = $form->getData();
 
             $legacyEnvironment = $environment->getEnvironment();
-
             $currentUser = $legacyEnvironment->getCurrentUser();
-
             $roomManager = $legacyEnvironment->getRoomManager();
             $roomItem = $roomManager->getItem($roomId);
 
-            $moderatorEmailAdresses = array();
-            $moderatorList = $roomItem->getModeratorList();
-            $moderatorUserItem = $moderatorList->getFirst();
-            while ($moderatorUserItem) {
-                $moderatorEmailAdresses[$moderatorUserItem->getEmail()] = $moderatorUserItem->getFullname();
-                $moderatorUserItem = $moderatorList->getNext();
-            }
+            $moderationRecipients = RecipientFactory::createModerationRecipients($roomItem);
 
-            $message = (new \Swift_Message())
-                ->setSubject($data['subject'])
-                ->setFrom(array($currentUser->getEmail() => $currentUser->getFullname()))
-                ->setTo($moderatorEmailAdresses)
-                ->setBody($data['message'])
-            ;
-
-            $message->setCc(array($currentUser->getEmail() => $currentUser->getFullname()));
-
-            $this->mailer->send($message);
+            $this->mailer->sendMultipleRaw(
+                $data['subject'],
+                $data['message'],
+                $moderationRecipients,
+                $currentUser->getFullName(),
+                [$currentUser->getEmail()]
+            );
 
             return new JsonResponse([
                 'message' => $translator->trans('message was send'),
@@ -852,27 +845,11 @@ class RoomController extends AbstractController
                     }
 
                     // mail to moderators
-                    $message = (new \Swift_Message())
-                        ->setFrom([$this->getParameter('commsy.email.from') => $roomItem->getContextItem()->getTitle()])
-                        ->setReplyTo([$newUser->getEmail() => $newUser->getFullName()]);
-
-                    $userManager = $legacyEnvironment->getUserManager();
-                    $userManager->resetLimits();
-                    $userManager->setModeratorLimit();
-                    $userManager->setContextLimit($roomItem->getItemID());
-                    $userManager->select();
-
-                    $moderatorList = $userManager->get();
-                    $moderator = $moderatorList->getFirst();
-                    $moderators = '';
-                    while ($moderator) {
-                        if ($moderator->getAccountWantMail() == 'yes') {
-                            $message->addTo($moderator->getEmail(), $moderator->getFullname());
-                            $moderators .= $moderator->getFullname() . "\n";
-                        }
-
-                        $moderator = $moderatorList->getNext();
-                    }
+                    $moderatorRecipients = RecipientFactory::createModerationRecipients(
+                        $roomItem, function ($moderator) {
+                        /** @var cs_user_item $moderator */
+                        return $moderator->getAccountWantMail() == 'yes';
+                    });
 
                     // language
                     $language = $roomItem->getLanguage();
@@ -885,11 +862,9 @@ class RoomController extends AbstractController
 
                     $translator = $legacyEnvironment->getTranslationObject();
 
-                    if ($message->getTo()) {
+                    if (!empty($moderatorRecipients)) {
                         $savedLanguage = $translator->getSelectedLanguage();
                         $translator->setSelectedLanguage($language);
-
-                        $message->setSubject($translator->getMessage('USER_JOIN_CONTEXT_MAIL_SUBJECT', $newUser->getFullname(), $roomItem->getTitle()));
 
                         $body = $translator->getMessage('MAIL_AUTO', $translator->getDateInLang(date("Y-m-d H:i:s")), $translator->getTimeInLang(date("Y-m-d H:i:s")));
                         $body .= "\n\n";
@@ -918,6 +893,11 @@ class RoomController extends AbstractController
                             $body .= "\n\n";
                         }
 
+                        $moderators = '';
+                        foreach ($moderatorRecipients as $recipient) {
+                            $moderators .= $recipient->getFirstname() . ' ' . $recipient->getLastname() .  "\n";
+                        }
+
                         $body .= $translator->getMessage('MAIL_SEND_TO', $moderators);
                         $body .= "\n";
 
@@ -935,9 +915,18 @@ class RoomController extends AbstractController
                             ], UrlGeneratorInterface::ABSOLUTE_URL);
                         }
 
-                        $message->setBody($body, 'text/plain');
-
-                        $this->mailer->send($message);
+                        $subject = $translator->getMessage(
+                            'USER_JOIN_CONTEXT_MAIL_SUBJECT',
+                            $newUser->getFullname(),
+                            $roomItem->getTitle()
+                        );
+                        $this->mailer->sendMultipleRaw(
+                            $subject,
+                            $body,
+                            $moderatorRecipients,
+                            $roomItem->getContextItem()->getTitle(),
+                            [$newUser->getEmail()]
+                        );
 
                         $translator->setSelectedLanguage($savedLanguage);
                     }
@@ -990,14 +979,13 @@ class RoomController extends AbstractController
                         'roomId' => $roomItem->getItemID(),
                     ], UrlGeneratorInterface::ABSOLUTE_URL);
 
-                    $message = (new \Swift_Message())
-                        ->setSubject($subject)
-                        ->setBody($body, 'text/plain')
-                        ->setFrom([$this->getParameter('commsy.email.from') => $roomItem->getContextItem()->getTitle()])
-                        ->setReplyTo([$contactModerator->getEmail() => $contactModerator->getFullName()])
-                        ->setTo([$newUser->getEmail()]);
-
-                    $this->mailer->send($message);
+                    $this->mailer->sendRaw(
+                        $subject,
+                        $body,
+                        RecipientFactory::createRecipient($newUser),
+                        $roomItem->getContextItem()->getTitle(),
+                        [$contactModerator->getEmail()]
+                    );
 
                     $translator->setSelectedLanguage($savedLanguage);
                 }
