@@ -9,15 +9,24 @@
 namespace App\Mail;
 
 
+use App\Entity\Account;
+use cs_room_item;
+use cs_user_item;
+
 class RecipientFactory
 {
-    public static function createModerationRecipients(\cs_room_item $room, callable $callback = null)
+    /**
+     * @param cs_room_item $room
+     * @param callable|null $callback
+     * @return array
+     */
+    public static function createModerationRecipients(cs_room_item $room, callable $callback = null): array
     {
         $recipients = [];
 
         $moderators = $room->getModeratorList();
         if (!$moderators->isEmpty()) {
-            /** @var \cs_user_item $moderator */
+            /** @var cs_user_item $moderator */
             $moderator = $moderators->getFirst();
 
             while ($moderator) {
@@ -36,7 +45,11 @@ class RecipientFactory
         return $recipients;
     }
 
-    public static function createRecipient(\cs_user_item $user): Recipient
+    /**
+     * @param cs_user_item $user
+     * @return Recipient
+     */
+    public static function createRecipient(cs_user_item $user): Recipient
     {
         $recipient = new Recipient();
         $recipient->setFirstname($user->getFirstname());
@@ -55,6 +68,44 @@ class RecipientFactory
         }
 
         $recipient->setLanguage($language);
+
+        return $recipient;
+    }
+
+    /**
+     * @param string $email
+     * @param string $firstname
+     * @param string $lastname
+     * @param string $language
+     * @return Recipient
+     */
+    public static function createFromRaw(
+        string $email,
+        string $firstname = '',
+        string $lastname = '',
+        string $language = 'de'
+    ): Recipient {
+        $recipient = new Recipient();
+        $recipient->setFirstname($firstname);
+        $recipient->setLastname($lastname);
+        $recipient->setEmail($email);
+        $recipient->setLanguage($language);
+
+        return $recipient;
+    }
+
+    /**
+     * @param Account $account
+     * @return Recipient
+     */
+    public static function createFromAccount(
+        Account $account
+    ): Recipient {
+        $recipient = new Recipient();
+        $recipient->setFirstname($account->getFirstname());
+        $recipient->setLastname($account->getLastname());
+        $recipient->setEmail($account->getEmail());
+        $recipient->setLanguage($account->getLanguage());
 
         return $recipient;
     }
